==> statistics/DESeq2/03_NCI-R25-analysisdata.Rmd <==
---
title: Create analysis ready data sets
jupyter:
  jupytext:
    text_representation:
      extension: .Rmd
      format_name: rmarkdown
      format_version: '1.2'
      jupytext_version: 1.11.3
  kernelspec:
    display_name: R
    language: R
    name: ir
---


## Import metadata

```{r}
library(tidyverse)
library(DESeq2)
```

```{r}
procdir <- normalizePath("./proc/")
countfile <- file.path(procdir, "PRJNA668393-countdata.RData")

countdata <- readRDS(countfile)

tools::md5sum(countfile)
```

```{r}
countdata %>% dim
```

```{r}
countdata %>% head(3)
```

## Import meta data

```{r}
procdir <- normalizePath("./proc/")

metafile <- file.path(procdir, "PRJNA668393-metadata.RDS")

mtdata <- readRDS(metafile)

tools::md5sum(metafile)
```

```{r}
mtdata %>% dim
```

## Create STAR count matrix


The columns names are of the form "SRR12804464_ReadsPerGene.out.tab". This function strips the "_ReadsPerGene.out.tab" suffix

```{r}
cleancolname <- function(x){
    stringr::str_remove(x, "_ReadsPerGene.out.tab")
}
cleancolname("SRR12804464_ReadsPerGene.out.tab")
```

STAR requires a count matrix with the genes across the rows and libraries across the columns. The first four rows will need to removed. The gene column will also need to be removed and assigned as row names

```{r}
### Labels for first four rows of STAR count files
starlabs <- c("N_ambiguous", "N_multimapping", "N_noFeature", "N_unmapped")

countdata %>% 
    dplyr::filter(!gene %in% starlabs) %>%
        data.frame() %>% 
            tibble::column_to_rownames("gene") %>%
                dplyr::rename_with(cleancolname) %>%
                    as.matrix -> 
                        starcountData
    
```

Inspect the dimension of the matrix

```{r}
starcountData %>% dim
```

Inspect the matrix

```{r}
starcountData[1:3,1:3]
```

## Create STAR column data


In this step we create the DESeq2 "column" data set from the metadata. The rows of this data object have to coincide with the columns of the count matrix.

The DESeq2 package expects the experimental variables (in this case treatment and cell_line) to be "factors". We add two new variables (trt and cline) as factors. DESeq2 does not like the "-" characters in the levels of its factors (e.g., as in the "anti-PD-1" level). We will remove these

```{r}
tibble::tibble(Run = colnames(starcountData)) %>%
    dplyr::full_join(mtdata, by = "Run") %>%
        dplyr::mutate(trt = factor(gsub("-", "",treatment), levels = c("isotype", "antiPD1")), 
                      cline = factor(cell_line, levels = c("Parental", "UV2"))) -> 
            starcolData
```

```{r}
starcolData
```

```{r}

```

```{r}

```

### Create DESeq2 data object


IMPORTANT: The rows of the STAR column data set have to line up with the columns of the STAR count data matrix. Be sure to check this very carefully. Following a merger, it is possible that the rows are reordered. 

```{r}
identical(starcolData[["Run"]], colnames(starcountData))
```

We can now construct a data object for use by the DESeq2 package

```{r}
PRJNA668393ddsData <- DESeq2::DESeqDataSetFromMatrix(countData = starcountData, colData = starcolData, design = ~1)
```

A quick inspection of this data object

```{r}
PRJNA668393ddsData
```

## Create a  data object for QCing the counts


### Tranpose count data


In the count data object, the genes are across the rows and the libraries are across the columns. To create a QC data set, the object will need to be transposed

```{r}
countdata %>% 
    tidyr::gather(fname, value, -gene) %>% 
        tidyr::spread(gene, value) %>% 
            head(3)
```

Add the total count for each library as a new column (depth) and drop gene counts

```{r}
countdata %>% 
    tidyr::gather(fname, value, -gene) %>% 
        tidyr::spread(gene, value) %>% 
            dplyr::rowwise() %>%
                dplyr::mutate(depth = sum(c_across(!contains(c("fname"))))) %>% 
                    dplyr::select(c(fname, all_of(starlabs), depth)) %>%
                        head(3)
```

Add the number of reads mapped to genes as a new column (N_mapped)

```{r}
countdata %>% 
    tidyr::gather(fname, value, -gene) %>% 
        tidyr::spread(gene, value) %>% 
            dplyr::rowwise() %>%
                dplyr::mutate(depth = sum(c_across(!contains(c("fname"))))) %>% 
                    dplyr::select(c(fname, all_of(starlabs), depth)) %>%
                        dplyr::mutate(N_mapped = depth - (N_ambiguous + N_multimapping + N_noFeature + N_unmapped)) %>%
                            head(3)
```

This function adds three QC statistics to the transposed count matrix: 1. proportion of mapped reads (defined as number of mapped reads over depth count), proportion of reads mapped to regions with no gene annotation, and proportion of unqiuely mapped reads

```{r}
mapstat <- function(cntdf) {
    cntdf %>%
        dplyr::mutate(prop.mapped =  N_mapped / depth) %>%
            dplyr::mutate(prop.nofeat = N_noFeature / depth) %>%
                dplyr::mutate(prop.unique = (N_mapped + N_noFeature) / depth)
}
```

Now, we add these QC statistics to the transposed count matrix

```{r}
countdata %>% 
    tidyr::gather(fname, value, -gene) %>% 
        tidyr::spread(gene, value) %>% 
            dplyr::rowwise() %>%
                dplyr::mutate(depth = sum(c_across(!contains(c("fname"))))) %>% 
                    dplyr::select(c(fname, all_of(starlabs), depth)) %>%
                        dplyr::mutate(N_mapped = depth - (N_ambiguous + N_multimapping + N_noFeature + N_unmapped))%>%
                            mapstat() %>%
                                head(3)
```

Finally merge in the meta data

```{r}
countdata %>% 
    tidyr::gather(fname, value, -gene) %>% 
        tidyr::spread(gene, value) %>% 
            dplyr::rowwise() %>%
                dplyr::mutate(depth = sum(c_across(!contains(c("fname"))))) %>% 
                    dplyr::select(c(fname, all_of(starlabs), depth)) %>%
                        dplyr::mutate(N_mapped = depth - (N_ambiguous + N_multimapping + N_noFeature + N_unmapped)) %>%
                            mapstat() %>%
                                dplyr::mutate(Run = cleancolname(fname)) %>%
                                    dplyr::full_join(mtdata, by = "Run") ->
                                          PRJNA668393qcData
PRJNA668393qcData %>% dim
PRJNA668393qcData %>% head(3)
```

```{r}
outdir <- normalizePath("./proc/")

qcfile <- file.path(outdir, "PRJNA668393-qcdata.RDS")
ddsfile <- file.path(outdir, "PRJNA668393-ddsdata.RDS")

saveRDS(PRJNA668393qcData, file = qcfile)
saveRDS(PRJNA668393ddsData, file = ddsfile)

tools::md5sum(qcfile)
tools::md5sum(ddsfile)
```

```{r}
sessionInfo()
```
